feat: split auth and non-auth app
