+++ src/configuration/navigation/Styles.js
@@ -1,0 +1,21 @@
+import { Colors, Fonts } from '../../modules/shared/themes'
+
+export const style = {
+  tabBarBackgroundColor: Colors.blue,
+  navBarButtonColor: Colors.white,
+  tabBarButtonColor: Colors.white,
+  navBarTextColor: Colors.white,
+  tabBarSelectedButtonColor: Colors.darkblue,
+  navigationBarColor: Colors.blue,
+  navBarBackgroundColor: Colors.blue,
+  statusBarColor: Colors.bloodOrange,
+  tabFontFamily: Fonts.type.base
+}
+
+export const barsHidden = {
+  navBarHidden: true,
+  tabBarHidden: true,
+  statusBarHidden: true,
+  drawUnderNavBar: true,
+  drawUnderTabBar: true
+}
